\name{BaSTA-package}
\alias{BaSTA-package}
\alias{BaSTA}

\docType{package}
\title{
BaSTA: a package for parametric Bayesian estimation of age-specific survival for truncated and censored capture-recapture/recovery records.
}

\description{
This package estimates mortality with covariates from capture-recapture/recovery data in a Bayesian framework when many individuals are of unknown age. It includes tools for data checking, model diagnostics and outputs such as life-tables and plots.
}

\details{
\tabular{ll}{
Package: \tab BaSTA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-10\cr
License: \tab GNU General Public Licence\cr
LazyLoad: \tab yes\cr
}

This package estimates mortality parameters from capture-recapture data where there are individuals with unknown birth/death dates.
It is currently under development and cannot be expected to work perfectly yet!

}

\author{
Fernando Colchero \email{colchero@mpg.demogr.de}
Owen Jones \email{jones@mpg.demogr.de}
Maren Rebke \email{rebke@mpg.demogr.de}
}

\references{
Colchero and Clark 2011 Bayesian inference on age-specific survival from capture-recapture data for censored and truncated data. Journal of Animal Ecology. 

Colchero, Jones and Rebke. BaSTA - a package to estimate mortality parameters from incomplete mark-recapture data with covariates. Unpublished Manuscript.
}

