\name{basta.out}
\alias{basta.out}
\docType{data}
\title{Output from a BaSTA analysis on a simulated dataset}
\description{
This dataset is the output of a BaSTA analysis on the simulated dataset \link{code{basta.dat}}. The analysis consisted of four independent simulations ran in parallel each of which was ran for 20,000 iterations. The model chosen was Gompertz (\dQuote{\code{GO}}) with argument \code{shape} set to \dQuote{\code{simple}} and \code{covarStruct} set to \dQuote{\code{fused}}.   
}
\usage{basta.out}
\format{RData file, list, basta.}
\keyword{datasets}