\name{sim1Out}
\alias{sim1Out}
\docType{data}
\title{Output from a Bayesian Survival Trajectory Analysis (BaSTA) analysis on a simulated dataset}
\description{
This dataset is the output of a BaSTA analysis on the simulated dataset \code{\link{sim1}}. The analysis consisted of four independent simulations run in parallel. Each simulation was run for 20,000 iterations. The model chosen was Gompertz (\dQuote{\code{GO}}) with the \code{shape} argument set to \dQuote{\code{simple}} and \code{covarStruct} set to \dQuote{\code{fused}}.   
}
\usage{sim1Out}
\format{RData file, list, basta.}
\keyword{datasets}