\name{BaSTA-package}
\alias{BaSTA-package}
\alias{BaSTA}
\docType{package}
\title{
BaSTA: a package for parametric Bayesian estimation of age-specific survival for truncated and censored capture-recapture/recovery records.
}

\description{
This package estimates survival trajectories with covariates from capture-recapture/recovery data in a Bayesian framework when many individuals are of unknown age. It includes tools for data checking, model diagnostics, and outputs such as life-tables and plots.
}

\details{
\tabular{ll}{
Package: \tab BaSTA\cr
Type: \tab Package\cr
Version: \tab 1.9.5\cr
Date: \tab 2022-10-05\cr
License: \tab GNU General Public Licence\cr
LazyLoad: \tab yes\cr
}

This package estimates survival parameters from capture-recapture data where there are individuals with unknown birth/death dates.

}

\author{
Fernando Colchero \email{colchero@imada.sdu.dk}, Owen Jones \email{jones@biology.sdu.dk}, and Maren Rebke \email{maren.rebke@avitec-research.de}
}

\references{
Colchero, F. and J.S. Clark (2012) Bayesian inference on age-specific survival from capture-recapture data for censored and truncated data. \emph{Journal of Animal Ecology}. 81, 139-149.

Colchero, F., O.R Jones and M. Rebke. (2012) BaSTA: an R package for Bayesian estimation of age-specific survival from incomplete mark-recapture/recovery data with covariates. \emph{Methods in Ecology and Evolution}. 3, 466-470.
}

