\name{PlotMultiMort}
\alias{PlotMultiMort}

\title{
Plots mortality trajectories for multiple BaSTA objects.}
\description{
This function produces a plot that plots the mortality trajectories of multiple BaSTA models on the same axes.
}
\usage{
PlotMultiMort(modList)
}

\arguments{
  \item{modList}{A list of the names of basta objects.}
 }
\details{}

\author{
Owen R. Jones \email{jones@demogr.mpg.de}
}

\seealso{
\code{\link{basta}}
}
\examples{
#PlotMultiMort(c("Model1","Model2"))
}
